// packages/client/src/defs.d.ts
export declare type PrismaAppSyncOptionsType = {
    connectionString?: string;
    sanitize?: boolean;
    debug?: boolean;
    defaultPagination?: number | false;
    maxDepth?: number;
};
export declare type Options = Required<PrismaAppSyncOptionsType> & {
    modelsMapping: any;
};
export declare type InjectedConfig = {
    modelsMapping?: {
        [modelVariant: string]: string;
    };
    operations?: string;
};
export declare type Action = typeof Actions[keyof typeof Actions] | string;
export declare type ActionsAlias = typeof ActionsAliases[keyof typeof ActionsAliases] | 'custom' | null;
export declare type ActionsAliasStr = keyof typeof ActionsAliases;
export declare type Context = {
    action: Action;
    alias: ActionsAlias;
    model: string | null;
};
// ...
export declare type QueryParams = {
    type: GraphQLType;
    operation: string;
    context: Context;
    fields: string[];
    paths: string[];
    args: any;
    prismaArgs: PrismaArgs;
    authorization: Authorization;
    identity: Identity;
    headers: any;
};
export declare type Authorization = typeof Authorizations[keyof typeof Authorizations] | null;
// ...
export declare type QueryParamsCustom = QueryParams & {
    prismaClient: PrismaClient;
};
// ...
export declare type AfterHookParams = QueryParams & {
    prismaClient: PrismaClient;
    result: any | any[];
};
export declare type Reason = string | ((context: Context) => string);
export declare type Shield = {
    [matcher: string]: boolean | {
        rule: boolean | any;
        reason?: Reason;
    };
};
// ...
export declare type HooksParameters<HookType extends 'before' | 'after', Operations extends string, CustomResolvers extends string> = {
    [matcher in HooksParameter<HookType, Operations, CustomResolvers>]?: (props: HooksProps[HookType]) => HooksReturn[HookType];
};
export declare type Hooks<Operations extends string, CustomResolvers extends string> = HooksParameters<'before', Operations, CustomResolvers> | HooksParameters<'after', Operations, CustomResolvers>;
export declare type ShieldAuthorization = {
    canAccess: boolean;
    reason: Reason;
    prismaFilter: any;
    matcher: string;
    globPattern: string;
};
export declare type ResolveParams<Operations extends string, CustomResolvers extends string> = {
    event: AppsyncEvent;
    resolvers?: {
        [resolver in CustomResolvers]: ((props: QueryParamsCustom) => Promise<any>) | boolean;
    };
    shield?: (props: QueryParams) => Shield;
    hooks?: Hooks<Operations, CustomResolvers>;
};
export { PrismaClient };
export declare type PrismaArgs = {
    where?: any;
    data?: any;
    orderBy?: any;
    skip?: number;
    take?: number;
    skipDuplicates?: boolean;
    select?: any;
};
export declare type AppsyncEvent = {
    arguments: any;
    source: any;
    identity: Identity;
    request: any;
    info: {
        fieldName: string;
        parentTypeName: string;
        variables: any;
        selectionSetList: string[];
        selectionSetGraphQL: string;
    };
    prev: {
        result: any;
    };
    stash: any;
};
export declare type GraphQLType = 'Query' | 'Mutation' | 'Subscription';
// ...
export declare type Identity = API_KEY | AWS_LAMBDA | AWS_IAM | AMAZON_COGNITO_USER_POOLS | OPENID_CONNECT;
// ...
export declare const BatchActionsList: string[];
export declare const DebugTestingKey = "__prismaAppsync";